+++ NINATest/FITSTest.cs
@@ -141,8 +141,8 @@ namespace NINATest {
                 new FITSHeaderCard("EGAIN", metaData.Camera.ElectronsPerADU, "[e-/ADU] Electrons per A/D unit"),
                 new FITSHeaderCard("XPIXSZ", metaData.Camera.PixelSize, "[um] Pixel X axis size"),
                 new FITSHeaderCard("YPIXSZ", metaData.Camera.PixelSize, "[um] Pixel Y axis size"),
-                new FITSHeaderCard("SET-TEMP", metaData.Camera.SetPoint, "[C] CCD temperature setpoint"),
-                new FITSHeaderCard("CCD-TEMP", metaData.Camera.Temperature, "[C] CCD temperature"),
+                new FITSHeaderCard("SET-TEMP", metaData.Camera.SetPoint, "[degC] CCD temperature setpoint"),
+                new FITSHeaderCard("CCD-TEMP", metaData.Camera.Temperature, "[degC] CCD temperature"),
             };
 
             var sut = new FITS(new ushort[] { 1, 2 }, 1, 1);
@@ -255,8 +255,8 @@ namespace NINATest {
                 new FITSHeaderCard("FOCPOS", metaData.Focuser.Position, "[step] Focuser position"),
                 new FITSHeaderCard("FOCUSPOS", metaData.Focuser.Position, "[step] Focuser position"),
                 new FITSHeaderCard("FOCUSSZ", metaData.Focuser.StepSize, "[um] Focuser step size"),
-                new FITSHeaderCard("FOCTEMP", metaData.Focuser.Temperature, "[C] Focuser temperature"),
-                new FITSHeaderCard("FOCUSTEM", metaData.Focuser.Temperature, "[C] Focuser temperature"),
+                new FITSHeaderCard("FOCTEMP", metaData.Focuser.Temperature, "[degC] Focuser temperature"),
+                new FITSHeaderCard("FOCUSTEM", metaData.Focuser.Temperature, "[degC] Focuser temperature"),
             };
 
             var sut = new FITS(new ushort[] { 1, 2 }, 1, 1);
@@ -280,6 +280,46 @@ namespace NINATest {
                 new FITSHeaderCard("ROTATOR", metaData.Rotator.Position, "[deg] Rotator angle"),
                 new FITSHeaderCard("ROTATANG", metaData.Rotator.Position, "[deg] Rotator angle"),
                 new FITSHeaderCard("ROTSTPSZ", metaData.Rotator.StepSize, "[deg] Rotator step size"),
+            };
+
+            var sut = new FITS(new ushort[] { 1, 2 }, 1, 1);
+            sut.PopulateHeaderCards(metaData);
+
+            foreach (var expectedCard in expectedHeaderCards) {
+                sut.HeaderCards.First(x => x.Key == expectedCard.Key).Should().BeEquivalentTo(expectedCard);
+            }
+        }
+
+        [Test]
+        public void FITSWeatherDataMetaDataPopulated() {
+            var now = DateTime.Now;
+            var metaData = new ImageMetaData();
+            metaData.WeatherData.CloudCover = 99.11;
+            metaData.WeatherData.DewPoint = 18.91;
+            metaData.WeatherData.Humidity = 46.52;
+            metaData.WeatherData.Pressure = 1010.4;
+            metaData.WeatherData.SkyBrightness = 43;
+            metaData.WeatherData.SkyQuality = 17.84;
+            metaData.WeatherData.SkyTemperature = -42;
+            metaData.WeatherData.StarFWHM = 2.34;
+            metaData.WeatherData.Temperature = 17.2;
+            metaData.WeatherData.WindDirection = 284.23;
+            metaData.WeatherData.WindGust = 1.76;
+            metaData.WeatherData.WindSpeed = 0.54;
+
+            var expectedHeaderCards = new List<FITSHeaderCard>() {
+                new FITSHeaderCard("CLOUDCVR", metaData.WeatherData.CloudCover, "[percent] Cloud cover"),
+                new FITSHeaderCard("DEWPOINT", metaData.WeatherData.DewPoint, "[degC] Dew point"),
+                new FITSHeaderCard("HUMIDITY", metaData.WeatherData.Humidity, "[percent] Relative humidity"),
+                new FITSHeaderCard("PRESSURE", metaData.WeatherData.Pressure, "[hPa] Air pressure"),
+                new FITSHeaderCard("SKYBRGHT", metaData.WeatherData.SkyBrightness, "[lux] Sky brightness"),
+                new FITSHeaderCard("MPSAS", metaData.WeatherData.SkyQuality, "[mags/arcsec^2] Sky quality"),
+                new FITSHeaderCard("SKYTEMP", metaData.WeatherData.SkyTemperature, "[degC] Sky temperature"),
+                new FITSHeaderCard("STARFWHM", metaData.WeatherData.StarFWHM, "Star FWHM"),
+                new FITSHeaderCard("AMBTEMP", metaData.WeatherData.Temperature, "[degC] Ambient air temperature"),
+                new FITSHeaderCard("WINDDIR", metaData.WeatherData.WindDirection, "[deg] Wind direction: 0=N, 180=S, 90=E, 270=W"),
+                new FITSHeaderCard("WINDGUST", metaData.WeatherData.WindGust * 3.6, "[kph] Wind gust"),
+                new FITSHeaderCard("WINDSPD", metaData.WeatherData.WindSpeed * 3.6, "[kph] Wind speed"),
             };
 
             var sut = new FITS(new ushort[] { 1, 2 }, 1, 1);